Add manpages to the package
Create the manpages for the API of the library. This manpages are still rough
and there might still be some mistakes or missing features. However it should
enough to start to use the library. From now, any changes (adding, deletion or
modification) to the API or behavior of the library SHOULD lead to a change in
the manpages.