Improve texture loading
Improve the code of the texture manager. Texture structure now holds in main
memory the image data which is loaded into video memory when the GL texture ID
is requested for the first time. This makes the creation of texture independent
to the the creation of OpenGL context (corresponding usually to the creation of
the window). Loading and destroying textures are now thread-safe.